GN-4131: PR review
